fix: SQL chart dashboard tile loading state
