+++ packages/frontend/src/features/sqlRunner/hooks/useSqlChartResults.ts
@@ -1,81 +1,12 @@
 import {
     isErrorDetails,
-    SchedulerJobStatus,
     type ApiError,
     type ApiJobScheduledResponse,
-    type ApiSqlRunnerJobStatusResponse,
     type ResultRow,
 } from '@lightdash/common';
 import { useQuery } from '@tanstack/react-query';
 import { lightdashApi } from '../../../api';
-import { getSchedulerJobStatus } from '../../scheduler/hooks/useScheduler';
-
-// To be reused across all hooks that need to fetch SQL query results
-const getResultsStream = async (url: string | undefined) => {
-    try {
-        if (!url) {
-            throw new Error('No URL provided');
-        }
-        const response = await fetch(url, {
-            method: 'GET',
-            headers: {
-                Accept: 'application/json',
-            },
-        });
-        const rb = response.body;
-        const reader = rb?.getReader();
-
-        const stream = new ReadableStream({
-            start(controller) {
-                function push() {
-                    void reader?.read().then(({ done, value }) => {
-                        if (done) {
-                            // Close the stream
-                            controller.close();
-                            return;
-                        }
-                        // Enqueue the next data chunk into our target stream
-                        controller.enqueue(value);
-
-                        push();
-                    });
-                }
-
-                push();
-            },
-        });
-
-        const responseStream = new Response(stream, {
-            headers: { 'Content-Type': 'application/json' },
-        });
-        const result = await responseStream.text();
-
-        // Split the JSON strings by newline
-        const jsonStrings = result.trim().split('\n');
-        const jsonObjects: ResultRow[] = jsonStrings
-            .map((jsonString) => {
-                try {
-                    return JSON.parse(jsonString);
-                } catch (e) {
-                    throw new Error('Error parsing JSON');
-                }
-            })
-            .filter((obj) => obj !== null);
-
-        return jsonObjects;
-    } catch (e) {
-        // convert error to ApiError
-        throw <ApiError>{
-            status: 'error',
-            error: {
-                name: 'Error',
-                statusCode: 500,
-                message: e.message,
-                data: {},
-            },
-        };
-    }
-};
+import { getResultsFromStream, getSqlRunnerCompleteJob } from './requestUtils';
 
 const getSqlChartResults = async ({
     projectUuid,
@@ -91,47 +22,12 @@ const getSqlChartResults = async ({
             body: undefined,
         },
     );
-
-    // recursively check job status until it's complete
-    async function getCompleteJob(): Promise<
-        ApiSqlRunnerJobStatusResponse['results']
-    > {
-        const job = await getSchedulerJobStatus<
-            ApiSqlRunnerJobStatusResponse['results']
-        >(scheduledJob.jobId);
-        if (
-            job.status === SchedulerJobStatus.SCHEDULED ||
-            job.status === SchedulerJobStatus.STARTED
-        ) {
-            return new Promise((resolve) => {
-                setTimeout(async () => {
-                    resolve(await getCompleteJob());
-                }, 1000);
-            });
-        }
-        if (job.status === SchedulerJobStatus.COMPLETED) {
-            return job;
-        } else {
-            throw <ApiError>{
-                status: 'error',
-                error: {
-                    name: 'Error',
-                    statusCode: 500,
-                    message: isErrorDetails(job.details)
-                        ? job.details.error
-                        : 'Job failed',
-                    data: {},
-                },
-            };
-        }
-    }
-
-    const job = await getCompleteJob();
+    const job = await getSqlRunnerCompleteJob(scheduledJob.jobId);
     const url =
         job?.details && !isErrorDetails(job.details)
             ? job.details.fileUrl
             : undefined;
-    return getResultsStream(url);
+    return getResultsFromStream(url);
 };
 
 export const useSqlChartResults = (
